move connection related functions to seperate file
